Twitter Sentiment Analysis
Used ‘Trump’ keyword to collect tweets to perform the sentiment analysis. 
Analyzed the ratio of positive/negative/stop word/others to the whole word count.
Observed the sum of all positive/negative words to say whether the general sentiment is negative or positive.

For each word that is not a “stop word” (and, the, a, is, as, ...), assigning a value +1 for positive sentiment or a value -1 for negative sentiment. Found a list of “stop words”, and a list of “negative” words online. For the words that are not in positive/negative/stop words, counting them as ‘others’.