Analysis using food recipe data
The objective of this project is to analyze food recipe data. This data is available at food.com. I intent to cover the below in this project scope:-
1. The average time of preparation of recipes.
2. Over what period was data collected and reviewed.
3. How are the variables correlated to each other.

This is going to help determine what are the factors that lead to successful/unsuccessful recipes. Looking at the above inferences, a blogger would know what is the average preparation time for recipes, how long does it take to be reviewed after submission and how are reviews correlated to year of submission.